Clean up netscraper package
* there is not Netscraper interface
* filters are available via options + extras
* renamed some methods for brevity
